requirements.txt
npm:
axios

pip:
django-cors-headers
